feat: 레포지토리 클래스 이름 수정 #28
